+++ lib/fakturoid/client.rb
@@ -2,6 +2,7 @@
 
 require 'fakturoid/client/account'
 require 'fakturoid/client/bank_account'
+require 'fakturoid/client/number_format'
 require 'fakturoid/client/user'
 require 'fakturoid/client/subject'
 require 'fakturoid/client/invoice'